Fix simulation order & add debug
DFF new state was evaluated in the "eval_output" function, which could
result in incorrect state if the input was not yet evaluated. Moving
this to the "eval_finish" pass fixes the problem.

We add a "debug_state" function to print the whole state of a
simulation (call is commented).